Add support for building in linux
- Exclude the VCPKG provided version of SDL2 in linux only and use standard rust-sdl2

In linux the VCPKG crate version of SDL2[vorbis] doesn't link correctly so the project doesn't build. As SDL2 libs can be trivially installed and linked via the package manager in linux I've added dependency config in Cargo.toml to just use normal dependencies for rust-sdl2 and added config to exclude the VCPKG version of the dependencies for linux so that it will still apply for macos and windows.